requirements.txt
python-dotenv==1.0.0
requests==2.31.0
openai>=1.12.0
schedule==1.2.0

